src/backend/requirements.txt
flask
torch
torchvision
pillow
pymongo